@@ -14,17 +14,18 @@ import (
 )
 
 // TestEmptyDirs models the following .stversions structure:
-// .stversions/
-// ├── keep1
-// │   └── file1
-// ├── keep2
-// │   └── keep21
-// │       └── keep22
-// │           └── file1
-// ├── remove1
-// └── remove2
-//     └── remove21
-//         └── remove22
+//
+//	.stversions/
+//	├── keep1
+//	│   └── file1
+//	├── keep2
+//	│   └── keep21
+//	│       └── keep22
+//	│           └── file1
+//	├── remove1
+//	└── remove2
+//		└── remove21
+//		    └── remove22
 func TestEmptyDirs(t *testing.T) {
 	var paths = []struct {
 		path   string
